+++ app/deployments/tasks.py
@@ -134,6 +134,23 @@ def handle_500_no_rows_error(timeseries_group, compare_text: str) -> bool:
     return False
 
 
+def handle_500_variable_actual_range_error(timeseries_group, compare_text: str) -> bool:
+    """ Did the request ask for a value that was outside of the real range of a value?
+
+    Returns True if handled.
+    """
+    if (
+        "Your query produced no matching results" in compare_text
+        and "is outside of the variable's actual_range" in compare_text
+    ):
+        logger.info(
+            f"{timeseries_group[0].dataset.name} with constraints {timeseries_group[0].constraints} had a constraint outside of normal range"
+        )
+        return True
+
+    return False
+
+
 def handle_500_time_range_error(timeseries_group, compare_text: str) -> bool:
     """ Did the request fall outside the range of times for the dataset
     
@@ -158,7 +175,7 @@ def handle_500_time_range_error(timeseries_group, compare_text: str) -> bool:
             try:
                 time = parse_time_string(potential_time)
                 times.append(time[0])
-            except DateParseError:
+            except (DateParseError, ValueError):
                 pass
         times.sort(reverse=True)
 
@@ -238,6 +255,11 @@ def handle_http_errors(timeseries_group, error: HTTPError) -> bool:
             if handle_500_time_range_error(timeseries_group, response_500.text):
                 return True
 
+            if handle_500_variable_actual_range_error(
+                timeseries_group, response_500.text
+            ):
+                return True
+
             logger.info(
                 f"500 error loading dataset {timeseries_group[0].dataset.name} with constraint {timeseries_group[0].constraints}: {error} ",
                 extra={
@@ -264,6 +286,9 @@ def handle_http_errors(timeseries_group, error: HTTPError) -> bool:
             return True
 
         if handle_500_time_range_error(timeseries_group, str(error)):
+            return True
+
+        if handle_500_variable_actual_range_error(timeseries_group, str(error)):
             return True
 
         logger.error(